requirements.txt
streamlit==1.22.0
requests==2.28.2
PyPDF2==3.0.1
Pillow==9.5.0
GitPython==3.1.31
markdown==3.4.3
